removed an unnecessary semicolon
git-svn-id: http://iterm2.googlecode.com/svn/branches/juhagman@707 5c829b59-20ba-c297-566c-b9ce016b2d6a

+++ Headers/iTerm/PTYTabView.h
@@ -52,7 +52,7 @@
 // NSTabView methods overridden
 - (void) addTabViewItem: (NSTabViewItem *) aTabViewItem;
 - (void) removeTabViewItem: (NSTabViewItem *) aTabViewItem;
-- (void) insertTabViewItem: (NSTabViewItem *) tabViewItem atIndex: (int) theIndex;
+- (void) insertTabViewItem: (NSTabViewItem *) tabViewItem atIndex: (NSInteger) theIndex;
 
 // selects a tab from the contextual menu
 - (void) selectTab: (id) sender;
